audio: bad audio clipping with GST_FORMAT_TIME
When the buffer timestamp/duration is having a non finished
value such as 0.333.. or 0.666.., the buffer clipping might loose
a sample causing a tick in the audio playback.

Indeed the diff with segment boundaries should be scaled and rounded
with 'gst_util_uint64_scale_round' and not with 'gst_util_uint64_scale'
in GST_FORMAT_TIME case.
Previously it was rounded always up but for 0.333.. diff it should be
rounded down.

See additional unit test:
'test_buffer_clip_time_handles_rounding_third'