ci: Build an arm64 LAVA rootfs
This rootfs will be used to run fluster tests on a LAVA arm64 device.
It also builds a linux kernel image and modules to boot the LAVA device.

The target device is mt8195-cherry-tomato-r2 and the kernel
configuration is adapted for it, whereas the rootfs is generic for arm64
devices.

The rootfs build is based on the Mesa CI. The support for other
architectures has been stripped but the structure of the build scripts
still keep the possibility to easily add other architectures in the
future. Only unused code as been removed.

2 debian dockers are built: 1 arm64 for building the rootfs and kernel,
and 1 amd64 that will be used later to prepare the artifacts, run the
Lava job and wait for results.