mpegtsmux: Lock incoming SCTE-35 to a pad
This is the continuation of the initial protection when dealing with SCTE-35
which are forwarded from an upstream mpeg-ts stream.

The problem is that the various (replicated) SCTE-35 events would be travelling
through different element chains which might have varying latency. So we could
end-up in a situation where those events would come in out of order (a later one
would come in through one sinkpad before the earlier one from another pad).

In order to fix that issue, when receiving SCTE sections from upstream we elect
a "prefered" pad for those sections, which will be the first one on which we
receive such events.