validate-scenario: Add non-blocking property to set-state
set-state currently waits for asynchronous state transitions to commit
before completing and executing any more actions.  While this is a
reasonable default, it makes certain tests impossible.

As an example, imagine a test with a pipeline containing playsink and
appsrc-push actions: It's perfectly reasonable to have the following
sequence of actions:

1. Set pipeline to PLAYING.
2. Push buffers with appsrc-push.

However, if the first action waits for the state to commit, the second
action can never start, therefore the state cannot preroll, and we get
deadlocked in a chicken-and-egg problem.

This patch adds a non-blocking property to the set-state action, which
allows to write validate tests for scenarios like this.