validateflow: Ignore stream property of stream-start by default
We don't want to record in expectation files any values that will
usually result in non-repeatable output.

For this reason, the "stream-id" property of the stream-start event is
already ignored by default.

However the "stream" property was not ignored by default. This patch
adds it to the default ignore list.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/7599>