ajasinkcombiner: clip audio to video pts and duration
If audio is clipped by audiobuffersplit before the aggregator and video
framerate changes, by the time we update audiobuffersplit properties it
is too late, it could already have clipped the next audio buffer for the
old framerate. The only place we can reliably clip audio is within the
aggregator after we receive the next video buffer.

An external audiobuffersplit also couldn't drop audio frames to align
with video pts.

TODO: What should we do if there are gaps in audio, should we insert
silence?