validate: reporter only abort with GST_VALIDATE_REPORTER_REPORT
In other words, no abort when GST_VALIDATE_REPORTER_KEEP.

When GST_VALIDATE_REPORTER_KEEP, gst_validate_runner_add_report is not
called, and inside that function is where the logic to check if
expected is executed.

Without this change, issues reported as expected-issue can abort the
execution if the report is kept by the reporter but not reported to
the runner (GST_VALIDATE_REPORTER_KEEP).