d3d12swapchainsink: Add properties and action signal for surface clear
Adding "clear-on-stop", "clear-color" properties to clear swapchain
surface. It can be useful if the sink is reused for multiple streams
and the application wants to clear swapchain when the sink is inactive
(i.e., null state). Also newly added "clear" action signal can be
used to clear swapchain anytime regardless of the state of sink.