libav: add allow-unaligned property to GstFFMpegVidDec
When a given video stream contains cropping information in-band, i.e. when
it is coded to certain dimensions but only a cropped should be displayed, FFmpeg
allows controlling the behavior via AVCodecContext::apply_cropping to
set whether FFmpeg shuold crop internally or provide the full coded pictures
plus cropping data in AVFrame's.

By default, apply_cropping is TRUE, and is also what we use in gst-libav.
However, the cropping done internally by FFmpeg is only done so far as it
preserves alignment. This can result in frames returned in a resolution
that is not the actual target cropped resolution. This behavior can also
be tuned by the AV_CODEC_FLAG_UNALIGNED, which is what is set when the
"allow-unaligned" property is set.

This property is se to FALSE by default to retain backwards
compatibility. Setting it to TRUE is relevant for testing decoder conformance,
when e.g. a checksum will calculated with the output.

Relevant stream: CROP_A_Panasonic_3.bit from ITU-T standard VVC streams.