baseparse: Allow controlling error-on-no-data behaviour
When no data was produced when EOS is received, the default behaviour is to post
an ERROR message. But this can be better handled downstream/externally by
elements which have a full view of the pipeline.

Add a property to allow such elements to demote the ERROR to a WARNING