ci: Remove gstreamer from debian and fedora images
This can interfere with correct testing of plugin loading in the
install step, and bloats the image massively.

Need newer Meson to avoid a dependency on qt6 lrelease, which pulls in
gstreamer indirectly.

We now also build our own opencv, to avoid pulling in 1GB of deps.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/8440>