sctp: flush data queue in cleanup
The pad task is waiting on the data_queue_pop call, which means that
when we try to shutdown sctpenc and join the task, it waits forever as
the data queue is never woken up. Set the queue to flushing and exit the
loop to ensure it shuts down cleanly.