pttdemux: Add a new action signal for requesting caps
Emitted when a new payload type is detected and a new src pad is about to
be created for it. This signal allows the application to provide accurate
caps (including SSRC and a-mid attributes) using information that might not
be available in ptdemux itself.

In case there was no request-caps signal handler or the handler returned NULL
caps, the request-pt-map signal is emitted.