va: Ensure valid max buffers in bufferpool config
We adjust min_buffers but not max, so we might end up with a case where the min
is greater than max, which causes an assert in
gst_buffer_pool_config_set_params().

Instead, if the bufferpool can't support the min buffers we need, let's just not
try to use it at all.