discoverer: fix duration being 0 at random with some file types
Some file types need additional parsing for their duration to be known.
The previous logic here was to go to PLAYING and sleep for a predefined
amount of time, but this can cause random failures on slow machines.

The new approach taken here is:
 * wait until we reach PAUSED
 * check the duration immediately and stop if it's known
 * otherwise, continue to PLAYING
 * wait for DURATION_CHANGED or EOS

This allows the discoverer to continue spinning in PLAYING and process
messages normally until its configured timeout. If the file has a
duration, it will stop when DURATION_CHANGED is received, otherwise
it will stop at EOS (eg. if it's an image).

Fixes: #4359