meson: Allow gstreamer-full to work with shared library plugins
While support for building gstreamer-full as a subproject should come by
letting the subprojects do
    build_target(target_type: '{shared,static}_library')
depending on if its doing a gstreamer-full build or not; this, while not
particularly useful, should still be supported for the edge-cases where
it is useful.

What this is missing right now is that the plugins will have to be
installed into libdir instead of libdir / 'gstreamer-1.0', as
gstreamer-full links against them like any other library.