autodetect: Avoid recursion if fpsdisplaysink ranks above autovideosink
When increasing the rank of fpsdisplaysink above autovideosink's (i.e.
setting GST_PLUGIN_FEATURE_RANK=fpsdisplaysink:99999999) the following
pipelines enter into an infinite recursion:

- gst-play-1.0 <path>
- gst-launch-1.0 videotestsrc ! videoconvert ! autovideosink

The solution is performed in `autodetect`, such way that if it is
trying to resolve an element which feature is the same feature as of
the one of its parent (in case it has a parent), then ignore such
candidate.

Illustrated diagram: the inner autovideosink (fps-displaysink-video-sink)
ignores selecting fpsdisplaysink as a candidate sink

----------------------------------------------------------------
| GstAutoVideoSink:autovideosink0                              |
|  ----------------------------------------------------------- |
|  | GstFpsDisplaySink:autovideosink0-actual-sink-fpsdisplay | |
|  |  ------------------------------------------------       | |
|  |  |  GstAutoVideosink:fps-displaysink-video-sink |       | |
|  |  |  -------------------                         |       | |
|  |  |  | candidate sink  |                         |       | |
|  |  |  -------------------                         |       | |
|  |  ------------------------------------------------       | |
|  ----------------------------------------------------------- |
----------------------------------------------------------------