nvcodec: Add precision timestamp SEI support
Add MISB ST 0604 precision timestamp support to NVIDIA hardware
encoders through SEI User Data Unregistered messages.

If a GstVideoMISPPrecisionTimestampMeta is present on the input buffer
the h.264 and h.265 encoders will write the timestamp to the SEI header.

This is enabled by default and can be disabled by using the new
`write-misb-precision-timestamp` property.