x264enc: Add precision timestamp SEI support
Add support for MISB ST 0604 precision timestamps through SEI
User Data Unregistered messages. This enables accurate timing
information to be embedded in H.264 streams.

If a GstVideoMISPPrecisionTimestampMeta is present on the input buffer
the h.264 and h.265 encoders will write the timestamp to the SEI header.

This is enabled by default and can be disabled by using the new
`write-misb-precision-timestamp` property.