qtdemux: handle LCEVC tracks
LCEVC enhancement data can be stored in the container as a separate track that
references a video track. This patch improves the muxer to attach LCEVC metadata
to video buffers if a separate LCEVC track references it.

Note that it is also possible to have LCEVC tracks in the container that don't
reference any video track. In this case, the demuxer will just expose a new src
pad with 'video/x-lvc1' caps, and push the LCEVC buffers there.