cmake: implement GStreamerMobile singleton library support
This commit implements the newly expanded GStreamerMobile singleton library.
Savings are pretty brutal e.g. on tutorial 5:

- Android.mk: 162 MB
- Current: 735.2 MB
- With the new singleton: 66.2 MB

Further savings may be achieved with element-based definitions.

This also implements a light cleanup, we can now use LINKER_LANGUAGE properly.

The new singleton support requires that on Android we pass
-DANDROID_USE_LEGACY_TOOLCHAIN_FILE=OFF as the LTO detection requires CMake 3.20
and NDK 22+ (we use 25). NDK 27+ is required to remove this workaround.