coreaudio: New src/sink elements for macOS
A new plugin built specifically for macOS, adding two new elements:
coreaudiosrc and coreaudiosink.

For end users, these function similarly to osxaudiosrc and sink but
should be much nicer to use thanks to the support for dynamic device
switching without stopping the pipeline, and much better error handling.

Internally, these elements have been written with a command-queue based
architecture which aims to reduce the possibility of deadlocks by
processing all commands in a single loop, making it easier to predict
the order of operations and hopefully allowing better maintainability in
the long run.

The device provider implementation is borrowed straight from osxaudio,
as it's using exactly the same APIs and there's no need to reinvent the
wheel.

Overall, aside from the device switching part, there isn't much to talk
about functionality-wise, but I hope that this will be a good baseline
for adding many useful capabilities in the future.