gst-video: Add new GstVideoEnhancementMeta API
This API replaces the old 'gstlcevcmeta.h' API from codecparsers library in bad,
making it obsolete. The new API will live in the video library from base so that
all video elements can easily handle LCEVC and other enhancement data. This will
be the case for muxer and demuxer elements, which will attach LCEVC metadata to
video buffers if the LCEVC enhancement data is stored as a separate track.