aacparse: use inner sample rate when producing ADTS
SBR (HE-AAC) in ADTS can only be implicitly signalled. When doing
implicit signalling we need to encode the sample rate of the AAC data
before any SBR adjustment.

Previously, when the source stream was using SBR with explicit
hierarchical signalling, aacparse was encoding the
extension_sampling_frequency, i.e. the sample rate *after* SBR as the
sample rate for the ADTS header. This would cause the decoder to use the
wrong sample rate for AAC LC content -- unless the content uses
downsampled SBR, where both fields have the same value.

This problem was encountered in Instagram in WebKit, but it would affect
any HE-AAC files that use explicit hierarchical signalling of
non-downsampled SBR.