decodebin3: remove deprecated stream collection
When decodebin3's collections rapidly expanding with collection only
adding new streams into previous ones, we need correctly label the
previous collection to deprecated and remove it when it's not the
currect output collection. When current output collection is deprecated,
we need switching to new output collection directly

Futher more, need ignore deprecated collection when finding collection
for stream, so that we can use the new updated collection.