v4l2videoenc: Set format on capture queue before encoder negotiate call
- On encoder capture queue, if the pixformat to be set is different
from what is already set by default, then call S_FMT with expected
pixformat before calling gst_v4l2_video_enc_negotiate(). Otherwise
G_CTRL call to get profile and level will fail for the expected format.
(e.g. G_CTRL for hevc profile will fail if the format set on driver
is AVC)

- If the pixformat changes from what is already set by default on
driver, then extracontrols which were already set are reset and needs
to be set again.

- S_FMT is called on CAPTURE port first and then later, on the OUTPUT
port The width and height values set using S_FMT call are ignored by the
driver and it returns default values.
Later, when S_FMT is called on output port, video driver internally
reconfigures the capture port with the desired width and height and
updates the sizeimage of the capture buffers.
But this updated sizeimage value is not updated in the Gstvideoinfo
of v4l2object of capture side and the bufferpool created during
decide_allocation remains with the older value.
This is causing assertions when the encoded output is of bigger size
than the gst buffers of v4l2bufferpool.
So to resolve this issue call enc_negotiate after set_format on
output port.