autodetect: Don't set the sync property of the kid with the default value
When autodetect (autoaudiosink/autovideosink) selects a sink element, the "sync"
property of this element is set to TRUE (default value of autodetect) even if this element
wants to disable the clock sync.

This commit allows to set the sync property of the kid only if it was set explicitly
on autodetect (autoaudiosink/autovideosink).