gl/egl: Request 8bit per channel by default
Only requesting one bit can result in non-optimal configs. Notably in
many cases 2101010 configs - with only 2bits for alpha - ends up being
selected, meaning that we get precission loss for alpha.

With RGBA (8888) being the default output format for the GL elements,
lets ensure we request configs that can represent that fully. For the
very unlikely cases that such a config is not supported, fall back to
the previous size of one.