uri: allow file URIs without authority part
Back in 2017, RFC 8089 defined the authority part of a file URI as
optional. This means that file URIs with one single leading slash
instead of the traditional three are valid.

This patch adds support for such URIs to gst_uri_get_location().

The discovery of the lack of support for these URIs ocurred as part of
the debugging in [^1], where it was found that Python 3.14 made
compatibility-breaking changes (arguably a regression) to the urljoin
function [^2].

[^1]: https://gitlab.freedesktop.org/gstreamer/gstreamer/-/issues/4757
[^2]: https://github.com/python/cpython/issues/125974